changed code structure for maintable files
