+++ SKGL 2.0.5.2/SKGL.csproj
@@ -51,8 +51,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Base\BaseConfiguration.cs" />
+    <Compile Include="Generate.cs" />
+    <Compile Include="SerialKeyConfiguration.cs" />
     <Compile Include="SKGL.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Validate.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="certificate.pfx" />
